test file split from original file
