+++ internal/auth/auth_test.go
@@ -3,26 +3,8 @@ package auth
 import (
 	"errors"
 	"net/http"
-	"strings"
 	"testing"
 )
-
-var ErrNoAuthHeaderIncluded = errors.New("no authorization header included")
-
-// GetAPIKey -
-func GetAPIKey(headers http.Header) (string, error) {
-	authHeader := headers.Get("Authorization")
-	if authHeader == "" {
-		return "", ErrNoAuthHeaderIncluded
-	}
-	splitAuth := strings.Split(authHeader, " ")
-	if len(splitAuth) < 2 || splitAuth[0] != "ApiKey" {
-		return "", errors.New("malformed authorization header")
-	}
-
-	return splitAuth[1], nil
-}
-
 
 func TestGetAPIKey(t *testing.T) {
 	tests := []struct {
@@ -54,7 +36,7 @@ func TestGetAPIKey(t *testing.T) {
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
 			got, err := GetAPIKey(tc.headers)
-			if got != tc.want || !errors.Is(err, tc.wantError) {
+			if got != tc.want || (err != nil && err.Error() != tc.wantError.Error()) {
 				t.Errorf("GetAPIKey() = %v, %v; want %v, %v", got, err, tc.want, tc.wantError)
 			}
 		})
